Fixed string functions not marked inline. Added number token reading.

+++ src/main.c
@@ -15,13 +15,13 @@ int main(int argc, char** argv)
 
 	struct string_ascii test = string_create_ascii("Hello World !\n");
 
-	int test_len = strlen(test.str);
+	int test_len = (int)strlen(test.str);
 
 	printf(test.str);
 	printf("Length = %d. test[%d] = %c\n", test_len, test_len, test.str[test_len]);
 
 	string_append_ascii(&test, "Adding another line !\n");
-	test_len = strlen(test.str);
+	test_len = (int)strlen(test.str);
 	printf(test.str);
 	printf("New Length = %d. test[%d] = %c\n", test_len, test_len, test.str[test_len]);
